pyproject.toml
[project]
name = "realtime-graphicframe"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "SeerLikeJazz", email = "liuzhaoxuchn@163.com" }
]
dependencies = [
    "pyqtgraph>=0.13.7",
    "pyserial>=3.5",
    "pyside6>=6.7.1",
    "matplotlib>=3.9.0",
    "numpy>=1.26.4",
    "scipy>=1.13.1",
    "pyautogui>=0.9.54",
    "pydirectinput>=1.0.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/realtime_graphicframe"]
